refactor: extract MeldungTransformationService for shared transformation logic and use logic in ProcessFromDirectory
