+++ src/test/java/org/miracum/streams/ume/obdstofhir/processor/Obdsv3ProcessorTest.java
@@ -81,6 +81,7 @@ import org.springframework.kafka.support.serializer.JsonSerializer;
       WeitereKlassifikationMapper.class,
       Obdsv2v3MapperConfig.class,
       Obdsv2v3MapperProperties.class,
+      MeldungTransformationService.class,
     })
 @EnableConfigurationProperties(value = {FhirProperties.class})
 public class Obdsv3ProcessorTest {
@@ -91,6 +92,7 @@ public class Obdsv3ProcessorTest {
 
   @Autowired private Obdsv3Processor processor;
   @Autowired private Obdsv3Deserializer obdsDeserializer;
+  @Autowired private MeldungTransformationService meldungTransformationService;
 
   @Value("classpath:obds3/*.xml")
   private Resource[] obdsResources;
@@ -320,7 +322,7 @@ public class Obdsv3ProcessorTest {
         buildMeldungExport(
             getResourceByName(inputXmlResourceName, obdsResources), "1", "12356789", "123", 1));
 
-    var result = processor.getTumorKonferenzmeldungen(meldungExportListV3);
+    var result = meldungTransformationService.getTumorKonferenzmeldungen(meldungExportListV3);
     assertThat(result.getFirst().getTumorkonferenz().getMeldeanlass().name())
         .isEqualTo(Meldeanlass.valueOf(expectedMeldeAnlass).name());
   }
@@ -393,7 +395,7 @@ public class Obdsv3ProcessorTest {
     meldungExportListV3.add(
         buildMeldungExport(
             getResourceByName("Test_SysT_2.xml", obdsResources), "3", "12356789", "125", 1));
-    var result = processor.getSystemtherapieMeldungen(meldungExportListV3);
+    var result = meldungTransformationService.getSystemtherapieMeldungen(meldungExportListV3);
 
     assertThat(result).as(description).isNotNull();
     assertThat(result.size()).as(description).isEqualTo(2);
